feat(cps1): Phase 10 -- sprite/BG field-layout + priority (corrects docs section 5/6)
1) cps1_oam_entry_t is now X,Y,tile,attr (was Y,tile,attr,X) with the real
attr bit layout: color 0-31 (attr widened uint8_t->uint16_t), X/Y flip,
and X/Y block-size nibbles for multi-tile sprites. cps1_ppu_render's
sprite base unit is now 16x16 (was 8x8, matching gfx(2)'s real MAME
layout), decomposed as a 2x2 grid of 8x8 sub-tiles -- same convention
cps1_bg.c already used for SCROLL2/3, not yet the real CPS1_GFX_LAYOUT_16X16
raw decode (deferred to Phase 11). Multi-tile block sprites iterate
(xblock+1)*(yblock+1) units; X/Y flip mirrors both the block arrangement
and each unit's own pixels, matching MAME's cps1_render_sprites DRAWSPRITE
loop (including the 512x512 internal coordinate wrap).

2) cps1_bg_cell_t is now literal code+attr (was tile_index/palette/an
invented "enabled" bit that doesn't exist on real hardware and collided
with attr's real bit 8 once attr became the literal word -- removed;
transparency is purely per-pixel via pen index 0 now, matching hardware).
Per-layer palette offset (+0x20/+0x40/+0x60 for SCROLL1/2/3) and X/Y flip
are implemented in cps1_blit8x8_indexed/cps1_blit_block_indexed (shared by
sprites and all three BG layers). CPS1_PALETTE_BANKS bumped 32->128 to
actually hold SCROLL3's reachable bank 0x7F.

3) Real bit-swizzled tilemap addressing (docs section 6) is implemented as
an exact forward/inverse bit-rearrangement pair (cps1_bg_swizzle_col_row_
to_offset / _offset_to_col_row) and wired into cps1_core.c's bus dispatch,
so cells[] stays simple logical row-major storage for the renderer while
the CPU-visible address space matches real hardware.

4) Replaced the unconditional bottom<middle<top compositor with
cps1_compositor_blend_priority: each BG pixel's cps1_blit8x8_indexed call
now also stamps a packed (priority_group, pen) byte per pixel; the new
compositor tests each CPS-B priority-mask register against the pen a BG
pixel of that group has, letting specific pens punch back over sprites
even though sprites otherwise draw on top of all BG layers -- the
plain-English hardware behavior from the confirmed MAME source comment,
deliberately NOT a port of MAME's internal tilemap set_transmask()/^0xffff
mechanism (docs section 9 Phase 10 plan says port the behavior, not the
mechanism). New CPS-B register block (0x800140) stores the 4 masks.

Documented, accepted architectural limitation: only SCROLL3-vs-sprite
priority is reproducible under cheat 8 on real hardware -- SCROLL1/2 become
LTDC hardware layers that blend final RGB565 with no pen-index knowledge,
so they can never punch over a sprite the way real CPS-1 silicon could.
cps1_core_run_frame_device_cost() only ever composites SCROLL3 against
sprites for this reason; cps1_core_run_frame() (host-correctness-only path)
composites all 3 layers for completeness testing.

Real bug found and fixed while implementing: cps1_resolve_base's Phase 9
"& (CPS1_GFXRAM_SIZE-1)" mask (not power-of-two) is unrelated to this
phase but was re-verified still correct via the full selftest suite below.

5) New/extended selftests, all passing, zero warnings:
- cps1-ppu-selftest: sprite footprint cull updated for the real 16x16 base
  (a coordinate that would have been wrongly culled under the old 8px
  assumption is now a regression guard), plus a new multi-tile block+flip
  render test (2x1 block, flip_x, asserts source-unit mirroring).
- cps1-bg-selftest: swizzle round-trip (all 3 layers, boundary-spanning
  col/row samples), per-layer palette-offset proof (with bank 1 poisoned
  to catch a regression to raw-color-as-bank), an asymmetric-pixel flip
  test, and a priority punch-through test (mask clear -> sprite wins, mask
  set -> BG pixel wins) via the new compositor.
- cps1-bg-bus-selftest / cps1-vdp-bus-selftest: updated for the new field
  widths/order (vdp_bus now targets sprite word 1 = Y under X,Y,tile,attr;
  bg_bus writes a value exercising color+flip+priority together).
- NEW cps1-priority-bus-selftest: proves the 68000 can reach the CPS-B
  priority-mask registers through the real dispatcher (the masks'
  behavioral EFFECT is proven separately in cps1-bg-selftest).

Verified:
- All 8 cps1 selftest binaries: OK, zero warnings, `make -f Makefile.cps1`
  clean build.
- retro-go-cps1 --engine=diff: same single frame-0 fb mismatch already
  documented as a Phase 9 diff-harness artifact (dual-engine shared-latch
  timing), cpu hash equal every frame; single-engine interpreter/
  recompiler runs converge to the identical RUNHASH.
- tools/m7_qemu_rig/run_cps1.sh: still links and runs on ARMv7-M. Honest
  cost update: device-cost path is now ~6.29M insn/frame (was ~1.85M pre-
  Phase-10), OVER the 60fps budget -- real 16x16 sprites (4x the pixels of
  the old 8x8 stub) plus per-pixel priority-meta stamping are genuinely
  more expensive than the under-scoped Phase 7-9 stub. cheat-8 saving is
  now 50.9% (was 69.7%). This is a real, expected cost increase from doing
  the actual field layout instead of a placeholder, not a regression to
  chase -- flagging honestly per this session's established practice, not
  masking it.

+++ linux/cps1/vdp_bus_selftest.c
@@ -17,7 +17,7 @@ int main(void)
 
     int16_t x = 0, y = 0;
     uint16_t tile = 0;
-    uint8_t attr = 0;
+    uint16_t attr = 0;
     cps1_core_oam_peek(0, &x, &y, &tile, &attr);
 
     printf("[cps1-vdp-bus-selftest] wram[0]=%u sprite0.y=%d "
